+++ src/Frontend/src/pages/mappers/panels/MapperDownloadPage.tsx
@@ -1,9 +1,7 @@
-import { Store } from "../../../utility/propertyStore"
 import { useState,  useEffect } from "preact/hooks";
 import { LoadProgress } from "../../../components/LoadProgress";
 import { MapperSelectionTable } from "./components/MapperSelectionTable";
 import { useAPI } from "../../../hooks/useAPI";
-import { MapperUpdate } from "pokeaclient";
 import { mapperFilesSignal, refreshMapperFiles } from "../../../Contexts/mapperFilesSignal";
 import { OpenMapperFolderButton } from "../../../components/OpenMapperFolderButton";
 import { Toasts } from "../../../notifications/ToastStore";
@@ -11,14 +9,14 @@ import { WideButton } from "../../../components/WideButton";
 import { Panel } from "@/components/Panel";
 import { Show } from "@preact/signals/utils";
 import { advancedModeSignal } from "@/Contexts/uiSettingsSignal";
+import { installMapper, MapperUpdate } from "@/utility/fetch";
 
 export function DownloadMapperPanel() {
-	const filesClient = Store.client.files;
 	const mapperFiles = mapperFilesSignal.value;
 	const [downloads, setDownloads] = useState<MapperUpdate[]>([]);
 	const [selectedDownloads, setSelectedDownloads] = useState<string[]>([]);
 	const downloadMappers = useAPI(
-		filesClient.downloadMapperUpdatesAsync, 
+		installMapper, 
 		(_, success) => {
 			if (success) {
 				refreshMapperFiles();
@@ -27,17 +25,17 @@ export function DownloadMapperPanel() {
 		}
 	);
 	useEffect(() => {
-		setDownloads(mapperFiles.updates.filter(mapper => !mapper.currentVersion) ?? []);
+		setDownloads(mapperFiles.updates.filter(mapper => !mapper.version) ?? []);
 		setSelectedDownloads([]);		
 	}, [mapperFiles.updates])
 
 	const handleDownload = () => {
-		const mappers = downloads.filter(x => selectedDownloads.includes(x.latestVersion.path));
-		downloadMappers.call(mappers);
+		const mappers = downloads.filter(x => selectedDownloads.includes(x.path));
+		downloadMappers.call(mappers.map(x => x.path));
 	}
 
 	const handleDownloadAll = () => {
-		downloadMappers.call(downloads);
+		downloadMappers.call(downloads.map(x => x.path));
 	}
 
 	if (downloadMappers.isLoading || mapperFiles.isLoading) {
@@ -61,6 +59,8 @@ export function DownloadMapperPanel() {
 					availableMappers={downloads}
 					selectedMappers={selectedDownloads}
 					onMapperSelection={setSelectedDownloads}
+					availableHeader="Version"
+					fallback={<strong>No mappers available for download.</strong>}
 				/>
 			</div>
 		</Panel>