+++ src/Frontend/src/pages/mappers/panels/components/MapperSelectionTable.tsx
@@ -1,23 +1,27 @@
-import { MapperUpdate, MapperVersion } from "pokeaclient";
 import { Dispatch, StateUpdater, useEffect, useState } from "preact/hooks";
 import { WideButton } from "../../../../components/WideButton";
+import { ComponentChildren } from "preact";
+import { MapperUpdate } from "@/utility/fetch";
 
 type MapperSelectionTableProps = {
 	onMapperSelection: Dispatch<StateUpdater<string[]>>,
 	availableMappers: MapperUpdate[],
 	selectedMappers: string[],
+	fallback: ComponentChildren,
+	installedHeader?: string
+	availableHeader?: string
 	onUpdateList?: () => void,
 }
 
 export function MapperSelectionTable(props: MapperSelectionTableProps) {
 	const { onMapperSelection, availableMappers, selectedMappers } = props;
-	const [allMappers, setAllMappers] = useState<MapperVersion[]>([]);
-	const [mappers, setMappers] = useState<MapperVersion[]>([]);
+	const [allMappers, setAllMappers] = useState<MapperUpdate[]>([]);
+	const [mappers, setMappers] = useState<MapperUpdate[]>([]);
 	const [filter, setFilter] = useState("");
 	useEffect(
 		() => {
-			setMappers(availableMappers.map(x => x.currentVersion ?? x.latestVersion).filter(x => !!x))
-			setAllMappers(availableMappers.map(x => x.currentVersion ?? x.latestVersion).filter(x => !!x))
+			setMappers(availableMappers)
+			setAllMappers(availableMappers)
 		},
 		[availableMappers, setMappers, setAllMappers]
 	);
@@ -59,7 +63,7 @@ export function MapperSelectionTable(props: MapperSelectionTableProps) {
 		return (
 			<>
 				<br />
-				<strong>No Mappers Found</strong>
+				{props.fallback}
 			</>
 		)
 	}
@@ -95,10 +99,16 @@ export function MapperSelectionTable(props: MapperSelectionTableProps) {
 							</label>
 						</th>
 						<th> Path </th>
+						{props.installedHeader &&
+							<th>{props.installedHeader}</th>
+						}
+						{props.availableHeader &&
+							<th>{props.availableHeader}</th>
+						}
 					</tr>
 				</thead>
 				<tbody>
-					{mappers.map((mapper, i) => {
+					{mappers.toSorted((a, b) => a.path > b.path ? 1 : -1).map((mapper, i) => {
 						return (
 							<tr key={i}>
 								<td >
@@ -113,6 +123,16 @@ export function MapperSelectionTable(props: MapperSelectionTableProps) {
 								<td>
 									{mapper.path}
 								</td>
+								{props.installedHeader &&
+									<td>
+										{mapper.version ?? "-"}
+									</td>
+								}
+								{props.availableHeader &&
+									<td>
+										{mapper.remote_version}
+									</td>
+								}
 							</tr>
 						)
 					})}