+++ src/PokeAByte.Domain/Interfaces/MapperDownloadDto.cs
@@ -1,0 +1,8 @@
+﻿namespace PokeAByte.Domain.Interfaces;
+
+public record MapperDownloadDto(
+    string RelativeXmlPath,
+    string XmlData,
+    string RelativeJsPath,
+    string JsData
+);