+++ src/PokeAByte.Web/Models/UpdateMemoryModel.cs
@@ -1,0 +1,3 @@
+namespace PokeAByte.Web.Models;
+
+public record UpdateMemoryModel(uint Address, byte[] Bytes);