+++ CHANGELOG.md
@@ -6,6 +6,18 @@
 [Version 0.9.0](#090)  
 [Version 0.8.1](#081-2024-11-09)  
   
+# 1.0
+
+## Breaking changes
+
+- Local mapper directories are no longer supported (a "mappers" folder next to the PokeAByte.Web.exe).
+
+## Breaking technical changes
+
+- `GET /mapper-service/get-mappers` now returns a different type of object.
+- `PUT /mapper` now expects a path, not an ID.
+- `PUT /mapper-service/change-mapper` now expects a path, not an ID.
+
 
 # 0.10.3